all: improve NT builds
- Added version information and icon
- Resolved warnings produced by MSVC 2019
- Fixed an issue whereby texture dimensions were not initialised before
  their use.

+++ inc/font.h
@@ -69,7 +69,7 @@ int FontPrintToRenderer(font_ctx *const ctx, const char *text,
  * \param w		Pointer to store expected width.
  * \param h		Pointer to store expected height.
  */
-void FontDrawSize(const size_t text_len, int *w, int *h);
+void FontDrawSize(const unsigned text_len, unsigned *w, unsigned *h);
 
 /**
  * Deletes font context.
